Work list styling
Putting a horizontal margin on the work experience list items and having the font size at 16 caused the title and right title to collide slightly on a larger device even.  Therefore, the font size was decreased to 14 and the margin taken out.

Also, I styled the text of the work experience points as the default styling wasn’t very attractive.